requirements.txt
flask==2.3.2
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
flask-cors

